19 GameStatus constant
