+++ packages/client/src/constants/game.ts
@@ -1,0 +1,6 @@
+export enum GameStatus {
+  Welcome = 'welcome',
+  Start = 'start',
+  Pause = 'pause',
+  End = 'end',
+}
